Fixes: * Made time-created field to be set by the poll service instead of the client.

+++ src/main/java/com/MTAPizza/Sympoll/pollmanagementservice/model/answer/Answer.java
@@ -22,5 +22,5 @@ public class Answer {
     private String answerContent;
 
     @Column(name = "num_of_votes")
-    private long numberOfVotes;
+    private int numberOfVotes;
 }
